third_party/blink/web_tests/external/wpt/css/css-values/calc-letter-spacing-expected.txt

This is a testharness.js-based test.
[FAIL] testing letter-spacing: calc(100%)
  assert_equals: calc(100%) should compute to 100% expected "100%" but got "20px"
Harness: the test ran to completion.
